Support querying with JWT tokens
A new example has been added, very similar to api_authz except
relevant information is conveyed via JWT tokens instead of as literals
in the policy. The example comes in the form of a second policy file
in the api_authz folder, and echo_server.py has been updated to work
with both policies.

Additionally, this patch updates the policies to be functionally
consistent with the online tutorials.