envoy_iptables: Add port whitelist to bypass envoy
This updates the script to have a new parameter `-w` which accepts a
comma separated list of ports. Inbound traffic for these ports will
be configured to bypass the envoy proxy by `RETURN`ing early in the
`ENVOY_IN_REDIRECT` chain.

By default there are no open ports, thus the change is backwards
compatible.

Any users of the OPA+Envoy integration that want to use a health check
on a separate port (using `--diagnostic-addr`) will likely need to
update their deployments to use this new flag.

Signed-off-by: Patrick East <east.patrick@gmail.com>