Kubernetes authorization webhook
Add example project demonstrating kubernetes webhook authorization using OPA.
Since most of the complexities with kubernetes authorization webhooks are not
about the actual authorization policies but rather about how to properly
configure kubernetes, this project uses kind to launch a local cluster for
testing.

To be followed up with a blog post going into the details of the configuration
and the caveats one needs to consider for this type of setup.

Signed-off-by: Anders Eknert <anders@eknert.com>